+++ examples/navigation-bottom-navigation-yew/src/main.rs
@@ -1,0 +1,19 @@
+#[cfg(all(feature = "csr", not(feature = "ssr")))]
+fn main() {
+    navigation_bottom_navigation_yew::hydrate();
+}
+
+#[cfg(all(feature = "ssr", not(feature = "csr")))]
+#[tokio::main]
+async fn main() {
+    println!("{}", navigation_bottom_navigation_yew::render_document());
+}
+
+#[cfg(all(feature = "csr", feature = "ssr"))]
+#[tokio::main]
+async fn main() {
+    println!("{}", navigation_bottom_navigation_yew::render_document());
+}
+
+#[cfg(all(not(feature = "csr"), not(feature = "ssr")))]
+fn main() {}